[REM-103] feat: add server blueprints and templates
- server/routes/blueprints/__init__.py: create blueprints for ASGI, auth, and Celery
- server/routes/__init__.py: initialize routes and register blueprints
- server/routes/blueprints/asgi/__init__.py: implement ASGI blueprint with status and control endpoints
- server/routes/blueprints/asgi/templates/status.html: add status template for ASGI server
- server/routes/blueprints/auth/templates/login.html: create login template for user authentication
- server/config.py: introduce StoreProcess dataclass for managing server processes
