+++ server/__init__.py
@@ -1,20 +1,18 @@
 """Module for manage application with System Status and System Information."""
 
-from dataclasses import dataclass
 from importlib import import_module
 from os import getenv
 from pathlib import Path
 
 import quart_flask_patch  # noqa: F401
 from flask_sqlalchemy import SQLAlchemy
-from quart import Quart, Response, make_response, render_template
+from quart import Quart
 from socketio import ASGIApp, AsyncServer
 
 instance_dir = Path(__file__).parent.resolve().joinpath("instance")
 template_dir = Path(__file__).parent.resolve().joinpath("templates")
 static_dir = Path(__file__).parent.resolve().joinpath("static")
 
-running_servers = {}
 
 app = Quart(
     __name__,
@@ -30,7 +28,6 @@ io = AsyncServer(
     ping_timeout=10,
 )
 
-import_module("server.logs", __package__)
 
 objects_config = {
     "development": "server.config.DevelopmentConfig",
@@ -39,33 +36,11 @@ objects_config = {
 }
 
 
-@dataclass
-class StoreProcess:
-    """Dataclass for storing process information."""
-
-    process_name: str
-    process_id: int
-    process_status: str
-    process_object: object
-
-
-@app.route("/dashboard")
-async def dashboard() -> Response:
-    """Render the dashboard template."""
-    return await make_response(await render_template("index.html", page="dashboard.html"))
-
-
-async def register_blueprint() -> None:
-    """Register the ASGI, worker, and beat blueprints."""
-    async with app.app_context():
-        from .blueprints import asgi_, auth_, beat_, worker_
-
-        for blueprint_ in [asgi_, worker_, beat_, auth_]:
-            app.register_blueprint(blueprint_)
-
-
 async def create_app() -> Quart:
     """Create and configure the Quart application."""
+    from server.routes import register_blueprint
+
+    import_module("server.logs", __package__)
     await register_blueprint()
     env_ambient = getenv("AMBIENT_CONFIG")
     ambient = objects_config[env_ambient]
