+++ server/routes/blueprints/asgi/__init__.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from billiard.context import Process
 from quart import Blueprint, Response, make_response, render_template
 
-from ... import StoreProcess, running_servers
+from server.config import StoreProcess, running_servers
 
 template_folder = Path(__file__).parent.resolve().joinpath("templates")
 static_folder = Path(__file__).parent.resolve().joinpath("static")
